Renamed toggle control so that it specifies that it's made specifically for actors.
