+++ src/sheets/actor/ActorEffectsTab.svelte
@@ -18,7 +18,7 @@
   import Notice from 'src/components/notice/Notice.svelte';
   import { declareLocation } from 'src/types/location-awareness.types';
   import ClassicControls from '../shared/ClassicControls.svelte';
-  import EffectToggleControl from 'src/components/item-list/controls/EffectToggleControl.svelte';
+  import ActorEffectToggleControl from 'src/components/item-list/controls/ActorEffectToggleControl.svelte';
 
   let context = getContext<Readable<ActorSheetContext>>('context');
 
@@ -38,7 +38,7 @@
 
     controls.push(
       {
-        component: EffectToggleControl,
+        component: ActorEffectToggleControl,
         props: ({ effect }) => ({
           effect: effect,
         }),
